chore: moved '/utilities/st-cuts' to '/'
